Some bugfixes for Mastery Victory
Fixed a mastery bug which caused the game to end when a cultural victory
would have been achieved.

Fixed a bug which caused the victory screen to show/hide the best rival
team at incorrect times.

Changed teamLaunchedShip in the victory screen to return true if the
spaceship is launched rather than landed. (The victory screen will need
further work to display the right information, but making the functions
do what their names suggest is a good policy.)